+++ example/shared/clientSystems/updateModelAttribute.lua
@@ -1,0 +1,12 @@
+local ReplicatedStorage = game:GetService("ReplicatedStorage")
+local Components = require(ReplicatedStorage.Game.components)
+
+local function updateModelAttribute(world)
+	for id, record in world:queryChanged(Components.Model) do
+		if record.new then
+			record.new.model:SetAttribute("clientEntityId", id)
+		end
+	end
+end
+
+return updateModelAttribute