+++ example/shared/start.lua
@@ -1,17 +1,27 @@
 local RunService = game:GetService("RunService")
-local Players = game:GetService("Players")
 local UserInputService = game:GetService("UserInputService")
 local ReplicatedStorage = game:GetService("ReplicatedStorage")
 local Packages = ReplicatedStorage.Packages
 local Matter = require(ReplicatedStorage.Lib.Matter)
 local Plasma = require(Packages.plasma)
 local HotReloader = require(Packages.rewire).HotReloader
+local components = require(script.Parent.components)
 
 local function start(container)
 	local world = Matter.World.new()
 	local state = {}
 
 	local debugger = Matter.Debugger.new(Plasma)
+
+	debugger.findInstanceFromEntity = function(id)
+		if not world:contains(id) then
+			return
+		end
+
+		local model = world:get(id, components.Model)
+
+		return model and model.model or nil
+	end
 
 	local loop = Matter.Loop.new(world, state, debugger:getWidgets())